matrices: move DeferredVector to matrixbase
